+++ src/main/kotlin/ru/yoomoney/tech/grafana/dsl/panels/alerting/Alert.kt
@@ -14,7 +14,8 @@ class Alert(
     private val frequency: Duration = 1.m,
     private val onNoData: AlertingState = Ok,
     private val onExecutionError: AlertingState = Alerting,
-    private val notificationIds: List<Any> = emptyList(),
+    private val notificationUids: List<String> = emptyList(),
+    private val notificationIds: List<Long> = emptyList(),
     private val conditions: AlertingConditions,
     private val pendingFor: Duration = 0.m
 ) : Json<JSONObject> {
@@ -28,7 +29,10 @@ class Alert(
         json["frequency"] = frequency.toString()
         json["noDataState"] = onNoData.asState()
         json["executionErrorState"] = onExecutionError.asState()
-        json["notifications"] = JSONArray(notificationIds.map { JSONObject().put("id", it) })
+        json["notifications"] = JSONArray(
+            notificationIds.map { JSONObject().put("id", it) } +
+                    notificationUids.map { JSONObject().put("uid", it) }
+        )
         json["conditions"] = conditions.toJson()
         json["for"] = pendingFor.toString()